requirements.txt
fastapi==0.115.11
uvicorn==0.41.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
twilio==9.10.2
google-generativeai
APScheduler
